+++ router/router_download.go
@@ -53,7 +53,7 @@ func getDownloadBackup(c *gin.Context) {
 
 	// Locate the backup on the local disk. We check this after token validation
 	// for security, but before attempting to open the file.
-	b, st, err := backup.LocateLocal(client, token.BackupUuid)
+	b, st, err := backup.LocateLocal(client, token.BackupUuid, token.ServerUuid)
 	if err != nil {
 		if errors.Is(err, os.ErrNotExist) {
 			c.AbortWithStatusJSON(http.StatusNotFound, gin.H{